Modify to add library capabilities alongside command line
Example usage:
```
from openapi2jsonschema import translator

result = translator.translate(schema="/path/to/swagger/swagger.yaml",output=None,root="SchemaRoot")
print(result)
```
